chore: fix two clippy lints from the audit batch
explicit_auto_deref in create_agent_task, unnecessary_cast on the
scheduler's attempt-number bind (the bulk count made it i64 already).

+++ crates/control-plane/src/store/scheduler.rs
@@ -435,7 +435,7 @@ impl Store {
         )
         .bind(&attempt_id)
         .bind(&task_id)
-        .bind(number as i64)
+        .bind(number)
         .bind(node_id)
         .bind(&lease)
         .bind(&ack_deadline)
